refactor: png => svg 로 변환
